centralized styling w/ lib pattern
