+++ aux_ui/lib/theme/aux_theme.dart
@@ -1,0 +1,8 @@
+
+library aux_theme;
+import 'package:flutter/material.dart';
+import 'dart:io' show Platform;
+
+part 'colors.dart';
+part 'text.dart';
+part 'scale_padding.dart';
