+++ aux_ui/lib/theme/scale_padding.dart
@@ -1,0 +1,31 @@
+part of 'aux_theme.dart';
+
+class SizeConfig {
+  static MediaQueryData _mediaQueryData;
+  static double screenWidth;
+  static double screenHeight;
+  static double blockSizeHorizontal;
+  static double blockSizeVertical;
+  
+  static double safeAreaHorizontal;
+  static double safeAreaVertical;
+  static double safeBlockHorizontal;
+  static double safeBlockVertical;
+  static EdgeInsets notchPadding;
+  
+  void init(BuildContext context) {
+    _mediaQueryData = MediaQuery.of(context);
+    screenWidth = _mediaQueryData.size.width;
+    screenHeight = _mediaQueryData.size.height;
+    blockSizeHorizontal = screenWidth / 100;
+    blockSizeVertical = screenHeight / 100;
+    
+    safeAreaHorizontal = _mediaQueryData.padding.left + 
+    _mediaQueryData.padding.right;
+    safeAreaVertical = _mediaQueryData.padding.top +
+    _mediaQueryData.padding.bottom;
+    safeBlockHorizontal = (screenWidth - safeAreaHorizontal) / 100;
+    safeBlockVertical = (screenHeight - safeAreaVertical) / 100;
+    notchPadding = ( Platform.isIOS ) ? EdgeInsets.only(top: SizeConfig.safeAreaVertical / 2, bottom: SizeConfig.safeAreaVertical / 2) : EdgeInsets.only(top: SizeConfig.safeAreaVertical); 
+  }
+}
