--- aux_ui/lib/theme/theme.dart
@@ -1,12 +1,0 @@
-// import 'package:aux_ui/theme/colors.dart';
-// import 'package:aux_ui/theme/text.dart';
-// import 'package:flutter/material.dart';
-
-// const auxTheme = ThemeData(
-//   primaryColor: auxPrimary,
-//   accentColor: auxAccent,
-//   fontFamily: 'Larsseit',
-//   textTheme: auxTextTheme, // TODO add secondary text theme for alt buttons?
-// );
-
-// ThemeData requires context
