+++ aux_ui/lib/widgets/aux_square_button.dart
@@ -1,2 +1,2 @@
 import 'package:flutter/material.dart';
-import 'package:aux_ui/theme/colors.dart';
+import 'package:aux_ui/theme/aux_theme.dart';
